refactor(nfa): code structure
