+++ pallets/non-fungible-assets/src/characteristics.rs
@@ -1,0 +1,6 @@
+pub mod bettor;
+
+pub trait AssetCharacteristic {
+  fn is_valid(&self) -> bool;
+}
+
